Throttle handling for Win OS performance queries
Rework how the OS performance queries are throttled in the Windows
OperatingSystemImpl. Replace the usage of clock() with _ftime() to allow
performance metrics to also be collected after 24.8 days when the
clock() method starts returning (clock_t)(-1).
The clock() method cannot reliably be used for long running processes, see
https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/clock?view=msvc-170

_ftime() provides time since epoch rather than the elapsed time since
the process started. For throttling how often performance metrics may be
retrived we can just as well compare to the previous epoch time.